Added customer location provider and set up map on customer tracking screen
