+++ lib/apps/customer_app/features/home/providers/booking.provider.dart
@@ -1,5 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:handees/apps/customer_app/features/tracker/providers/customer_location.provider.dart';
 import 'package:handees/apps/customer_app/services/booking_service.customer.dart';
 import 'package:handees/apps/customer_app/services/sockets/customer_socket.dart';
 import 'package:handees/shared/data/handees/job_category.dart';
@@ -9,21 +10,21 @@ final bookingProvider = StateNotifierProvider<BookingNotifier, BookingState>(
     (ref) => BookingNotifier(
         FirebaseAuth.instance,
         ref.watch(bookingServiceProvider),
-        Location.instance,
-        ref.watch(customerSocketProvider)));
+        ref.watch(customerSocketProvider),
+        ref));
 
 class BookingNotifier extends StateNotifier<BookingState> {
   final BookingService _bookingService;
   final FirebaseAuth _auth;
-  final Location _location;
   final CustomerSocket _socket;
+  final StateNotifierProviderRef<BookingNotifier, BookingState> _ref;
 
   late JobCategory _category;
   JobCategory get category => _category;
 
-  BookingNotifier(
-      this._auth, this._bookingService, this._location, this._socket)
+  BookingNotifier(this._auth, this._bookingService, this._socket, this._ref)
       : super(BookingState.idle) {
+    _ref.read(customerLocationProvider.notifier).initLocation();
     _socket.connect();
     _socket.onBookingOfferAccepted((event) {
       state = BookingState.inProgress;
@@ -40,27 +41,6 @@ class BookingNotifier extends StateNotifier<BookingState> {
     super.dispose();
   }
 
-  void init() async {
-    bool serviceEnabled;
-    PermissionStatus permissionGranted;
-
-    serviceEnabled = await _location.serviceEnabled();
-    if (!serviceEnabled) {
-      serviceEnabled = await _location.requestService();
-      if (!serviceEnabled) {
-        return;
-      }
-    }
-
-    permissionGranted = await _location.hasPermission();
-    if (permissionGranted == PermissionStatus.denied) {
-      permissionGranted = await _location.requestPermission();
-      if (permissionGranted != PermissionStatus.granted) {
-        return;
-      }
-    }
-  }
-
   Future<String> bookService({
     required JobCategory category,
   }) async {
@@ -68,14 +48,14 @@ class BookingNotifier extends StateNotifier<BookingState> {
     _category = category;
 
     final token = await _auth.currentUser!.getIdToken();
-    final location = await _location.getLocation();
+    final location = await Location.instance.getLocation();
 
     return await _bookingService.bookService(
       token: token,
-      // lat: location.latitude!,
-      // lon: location.longitude!,
-      lat: 6.548281268456966,
-      lon: 3.332248000980724,
+      lat: location.latitude!,
+      lon: location.longitude!,
+      // lat: 6.548281268456966,
+      // lon: 3.332248000980724,
       category: category,
     );
   }
